+++ scripts/wrap.lua
@@ -3413,7 +3413,7 @@ end
                     filters = {"stop()", "tick(timerId)"}
                 },
                 system = {
-                    filters = {"actionStart(action)", "actionStop(action)", "actionLoop(action)", "update()", "flush()"}
+                    filters = {"actionStart(action)", "actionStop(action)", "actionLoop(action)", "update()", "flush()", "inputText(text)"}
                 },
                 library = {}
             }